requirements.txt
torch==2.7.0
streamlit==1.45.1
python-telegram-bot==22.1
python-dotenv==1.1.0